+++ settings/footersettings.php
@@ -53,6 +53,12 @@ defined('MOODLE_INTERNAL') || die();
     $setting->set_updatedcallback('theme_reset_all_caches');
     $page->add($setting);
 
+    $name = 'theme_stream/twitterurl';
+    $title = get_string('twitterurl', 'theme_stream');
+    $setting = new admin_setting_configtext($name, $title, '' , 'https://www.twitter.com/mytwitterhandle');
+    $setting->set_updatedcallback('theme_reset_all_caches');
+    $page->add($setting);
+
     $name = 'theme_stream/youtubeurl';
     $title = get_string('youtubeurl', 'theme_stream');
     $setting = new admin_setting_configtext($name, $title, '' , 'https://www.youtube.com/myyoutubepage');